requirements.txt
python-chess>=1.10.0
playwright>=1.40.0
numpy>=1.24.0
